### 2022-03-05 #4222 @Kevin-Mattheus-Moerman
@baxmittens thanks for this submision, I am the AEiC for JOSS on call the coming week. I've briefly reviewed this submission and have just triggered an additional scope review by our editorial board. This is normal for submissions of this size (in terms of lines of code and functionality). This scope review should take about 2 weeks to complete.
In the mean time, perhaps you can check and address these potentially missing DOIs :point_up:, you can call `@editorialbot generate pdf` here to update the paper and you can check the references using `@editorialbot check references`.
