### 2022-03-22 #4222 @pdebuyl
Hi @kthyng may I deflect this? I am interested but got a lot of delay on other work items, including my JOSS editorial duties, since january and would like to focus on those in the short term.
