### 2022-03-24 #4222 @dmbates
I just glanced at the repository for this package and there are several issues that I think should be addressed before the paper is considered for publication in JOSS.  As it stands the repository is not in the standard form for a Julia package.
- I personally don't know what a VTU file is - from the example I assume the V stands for Voxel but that is about all I can figure out.  The purpose of the package should be explained more clearly and a reference to the definition of the VTU file format included.
- One very trivial issue is that the name of the repository should be VTUFileHandler.jl not VTUFileHandler.
- At present there is no `docs` directory and no inline documentation in the code. This means there will be no documentation available using the help system in Julia.
- The `include` statements in `src/VTUFileHandler.jl` are not in the standard form, using relative paths that can become unreliable
- It is more of a stylistic issue but burying directives like `import TranscodingStreams` and `import Dates` way down in the source code tree is unusual.  It is more common to put these in `src/VTUFileHandler.jl` - in most cases right after the declaration of the module name.

I would recommend that the authors first revise the package to follow the Julia package standards, add documentation for the exported functions and structs, and submit the package for inclusion in the general package registry, then reconsider the paper.  It is not a particularly onerous task to get a Julia package registered but it is not trivial either.  There are some resources such as the `PackageTemplates` package and the JuliaRegistrator bot but it still takes some effort the first time through.  There is a slack channel dedicated to questions on package registration.
