### 2022-03-25 #4222 @baxmittens
hello @dmbates thank you for reviewing and providing some initial feedback.

> I just glanced at the repository for this package and there are several issues that I think should be addressed before the paper is considered for publication in JOSS. As it stands the repository is not in the standard form for a Julia package.

I don't think my package it should be a standard julia package at this point. There is [WriteVTK.jl](https://github.com/jipolanco/WriteVTK.jl) and [ReadVTK.jl](https://github.com/trixi-framework/ReadVTK.jl). I did not really want to use those packages at the moment but I would consider it in the future. As I wrote in the paper, my project is mainly about automation of stochastic postprocessing and therefore the need to implement a basic algebra on a result file format.

> * I personally don't know what a VTU file is - from the example I assume the V stands for Voxel but that is about all I can figure out.  The purpose of the package should be explained more clearly and a reference to the definition of the VTU file format included.

I included that in the Readme.md

> * One very trivial issue is that the name of the repository should be VTUFileHandler.jl not VTUFileHandler.

Changed.

> * At present there is no `docs` directory and no inline documentation in the code. This means there will be no documentation available using the help system in Julia.

You're right. I will add some inline documentation.

> * The `include` statements in `src/VTUFileHandler.jl` are not in the standard form, using relative paths that can become unreliable

Could you please give me a short example how this would look like for a project with subfolders? I looked for it but could find any. I thought that something like `include(joinpath(".","VTUFileHandler","utils.jl"))` should work for Apple/Windows/Linux.

> * It is more of a stylistic issue but burying directives like `import TranscodingStreams` and `import Dates` way down in the source code tree is unusual.  It is more common to put these in `src/VTUFileHandler.jl` - in most cases right after the declaration of the module name.

I'll change this.

> I would recommend that the authors first revise the package to follow the Julia package standards, add documentation for the exported functions and structs, and submit the package for inclusion in the general package registry, then reconsider the paper. It is not a particularly onerous task to get a Julia package registered but it is not trivial either. There are some resources such as the `PackageTemplates` package and the JuliaRegistrator bot but it still takes some effort the first time through. There is a slack channel dedicated to questions on package registration.

I think I can fix these issues rather quickly. But as I said I do not plan to publish this as a official julia package at this stage of the project.

The VTUFileHandler is just one part of a stochastic toolbox for geoscience applications I am developing consisting of:

- OGSUQ.jl: Stochastic Toolbox for massive parallel computations of [OGS6 ](https://www.opengeosys.org/) simulations and automated stochastic postprocessing (mean, expval, var, sobol-indices,...).
- OGSInputfileHandler.jl: A basic julia implementation of [ogs6py](https://joss.theoj.org/papers/10.21105/joss.03673)
- AdaptiveSparseGrids.jl: A julia implementation of an adaptive hierarchical stochastic collocation method ([paper](https://arxiv.org/pdf/1509.01462.pdf)).

AdaptiveSparseGrids.jl and OGSUQ.jl could serve for two additional joss papers. My plan was to upload these projects later this year and subsequently submit two additional papers, here.
I'll think if this is all finished, I'll reconsider publishing the whole project as standard julia package.
