### 2022-03-29 #4222 @Datseris
Hello, thank you for the invitation but unfortunately I have to kindly decline, because the paper topic is too far away from my area of expertise and I am already dedicating some free time resources reviewing https://github.com/openjournals/joss-reviews/issues/4207

I wish you the best in finding alternative reviewers!
