### 2022-04-08 #4300 @jbytecode
@baxmittens  - at a first glance, it seems a citation for Julia is missing. could you please consider adding it?

@dmbates, @mkitti - could you please create your checklist by typing the editorialbot command given above?

Thank you all in advance.
