### 2022-04-09 #4300 @mkitti
The authors introduce VTUFileHandler.jl, a Julia language package for handling VTK unstructured grid data. The package also provides a number of mathematical operators reminiscent of an algebraic field over addition and multiplication.

1. Currently VTUFileHandler.jl is not a registered Julia package. It depends on another unregistered package located at https://github.com/baxmittens/XMLParser, which is a bespoke "leightweight" [spelling error] package.

Does the author intend to register one or both packages?

Given the the XMLParser module is bespoke, is it really meant to be an independent package? Could it exist as a module within VTUFileHandler.jl?

2. Since part of the package deals with file based input and output, it would be useful if this functionality could be registered file handler via [FileIO.jl](https://juliaio.github.io/FileIO.jl/stable/implementing/)

3. The registered package [`WriteVTK.jl`](https://github.com/jipolanco/WriteVTK.jl) appears to implement file loading and saving for `.vtu` files. A comparison to this package should be done. I would also appreciate if the author could explore interoperability with this package.

4. While the author asserts that this library "implements an algebra" I would like the author to defend this statement by addressing a few algebraic concepts.
    i) Are the operations associative?
    ii) Are the operations commutative?
    iii) Are there additive and multiplicative identities?
    iv) Are there additive and multiplicative inverses?
    v) Can multiplication be distributed over addition?
    vi) Over the set of `VTUFile`s do addition and multiplication form a [mathematical field](https://en.wikipedia.org/wiki/Field_(mathematics))?
    v) Over the set of `VTUFile`s and `Number`s do addition and multiplication form a multiplicative field?

5. The method `nbytes` is implemented as a utility within the package: https://github.com/baxmittens/VTUFileHandler.jl/blob/633c67c0d1f50c3d75903448b7800c8fa944d407/src/VTUFileHandler/vtuheader_utils.jl#L13-L17 . This appears to replicate the functionality of [`sizeof`](https://docs.julialang.org/en/v1/base/base/#Base.sizeof-Tuple{Type})

```julia
julia> nbytes(UInt8)
1

julia> sizeof(UInt32)
4

julia> sizeof(UInt64)
8

julia> sizeof(Float64)
8

julia> sizeof(Int32)
4

julia> sizeof(Int64)
8

julia> sizeof(UInt8)
1
```

Could the definition `const nbytes = sizeof` replace the current hardcoded version?

_To be continued..._
