### 2022-04-09 #4300 @baxmittens
> @baxmittens - at a first glance, it seems a citation for Julia is missing. could you please consider adding it?

Gladly did so.
