### 2022-04-09 #4300 @baxmittens
@mkitti - thank you for your comments. I would like to answer to them

> 1. Currently VTUFileHandler.jl is not a registered Julia package. It depends on another unregistered package located at https://github.com/baxmittens/XMLParser, which is a bespoke "leightweight" [spelling error] package.

Fixed the typo.

> Does the author intend to register one or both packages?

Not at this point but maybe later. I elaborate on this below

> Given the the XMLParser module is bespoke, is it really meant to be an independent package? Could it exist as a module within VTUFileHandler.jl?

I use it in other projects of mine too. For example for some simple html and OpenGeoSys input file manipulations. Thinking in term of functions, XMLParser.jl should be a stand-alone package since it has a distinctive function. I'm no big fan of mixing too many diffierent functionalities in one project since it complicates the expandability and the maintainbility.
What I would consider is replacing my XMLParser for a better one. The problem is, to my knowledge, there is no XMLParser available which is written purely in Julia. All projects I know (LightXML.jl , EzXML.jl) use [libxml2](http://www.xmlsoft.org/) as a basis, which I wanted to avoid.

And finally, even VTUFileHandler.jl is only part of a bigger project. As I wrote earlier (in Prereview), I implemented an adaptive sparse-grid colloction method ([see sparse grid theory paper](https://arxiv.org/abs/1509.01462)) which uses the VTUFileHandler and a broader stochastic / hpc framework which uses the adaptive sparse-grid. I could throw this all in one package but I thought it would be much better to distribute this big project in smaller packages according to their functionalities. I plan on publishing two consecutive papers describing the sparse-grid and the stochastic-hpc-framework

> 2. Since part of the package deals with file based input and output, it would be useful if this functionality could be registered file handler via [FileIO.jl](https://juliaio.github.io/FileIO.jl/stable/implementing/)
> 3. The registered package [`WriteVTK.jl`](https://github.com/jipolanco/WriteVTK.jl) appears to implement file loading and saving for `.vtu` files. A comparison to this package should be done. I would also appreciate if the author could explore interoperability with this package.

Before starting this project, I looked into both WriteVTK.jl and ReadVTK.jl and at that point both package were not really a good fit for what I was planning on doing. But as things progress I would definetly consider rebasing my stuff onto their projects.
That is a reason why I don't think VTUFileHanlder.jl should be a registered julia package at the moment.

> 4. While the author asserts that this library "implements an algebra" I would like the author to defend this statement by addressing a few algebraic concepts.
>    i) Are the operations associative?
>    ii) Are the operations commutative?
>    iii) Are there additive and multiplicative identities?
>    iv) Are there additive and multiplicative inverses?
>    v) Can multiplication be distributed over addition?
>    vi) Over the set of `VTUFile`s do addition and multiplication form a [mathematical field](https://en.wikipedia.org/wiki/Field_(mathematics))?
>    v) Over the set of `VTUFile`s and `Number`s do addition and multiplication form a multiplicative field?

As mentioned in the paper, if operators should be applied the objects have to share the same topology. If they do, all operators act *point-wise* on each corresponding pair of e.g. nodal coefficients. The operators itself keep their properties as they perform their action on a collection of scalars. Each nodal coefficient  $\hat{u}\in\mathbb{R}$ is an independet entity.

> 5. The method `nbytes` is implemented as a utility within the package: https://github.com/baxmittens/VTUFileHandler.jl/blob/633c67c0d1f50c3d75903448b7800c8fa944d407/src/VTUFileHandler/vtuheader_utils.jl#L13-L17 . This appears to replicate the functionality of [`sizeof`](https://docs.julialang.org/en/v1/base/base/#Base.sizeof-Tuple%7BType%7D)

Thank you. Changed all nbytes to sizeof. That makes a lot more sense...

Sorry for the long answer.
