### 2022-04-09 #4300 @jbytecode
Look likes there are lots of "my needs" instead of "community's needs", the software can be refactored in lights of the reviews.

The dependency XMLParser as in the declared in Project.toml is not a standard Julia package - I am sharing the same thoughts with the reviewer.

Altought it is not strictly needed, we always encourage authors to submit their packages so it fits well with the eco-system of the language and the environments.

At the end of the story, the software should solve a generalized problem with an easy and painless installation process.
