### 2022-04-09 #4300 @baxmittens
> Looks like there are lots of "my needs" instead of "community's needs", the software can be refactored in lights of the reviews.
>
> The dependency XMLParser as declared in the Project.toml is not a standard Julia package - I am sharing the same thoughts with the reviewer.
>
> Altought it is not strictly needed, we always encourage authors to submit their packages so it fits well with the eco-system of the language and the environments.
>
> At the end of the story, the software should solve a generalized problem with an easy and painless installation process.

Agree, this was a project that originally was created only in my needs and I was asked to open-source it because other people wanted to use it. But I do think it solves a generalized problem quite handy.
And I also think the installation process is painless since no package has dependencies which is quite nice if you think about hpc computations on heterogenous cluster architectures.
