### 2022-04-12 #4300 @mkitti
> Before starting this project, I looked into both WriteVTK.jl and ReadVTK.jl and at that point both package were not really a good fit for what I was planning on doing. But as things progress I would definetly consider rebasing my stuff onto their projects.
> That is a reason why I don't think VTUFileHandler.jl should be a registered julia package at the moment.

Regarding package registration, I do not see how dependence on the other packages or lack there of would prevent package registration. This can always be modified in subsequent vesions of the package. VTUFileHandler.jl is either a reusable library or a terminal application.
1) A reusable library should be in a registry of some sort to facilitate installation.
2) Alternatively, a terminal application should have a Manifest.toml for reproducibility as well as a DOI via a service like Zenodo.

Per the review criteria on https://joss.readthedocs.io/en/latest/review_criteria.html#installation-instructions:

> Where possible, software installation should be managed with a package manager. However, this criterion depends upon the maturity and availability of software packaging and distribution in the programming language being used. For example, Python packages should be pip installable, and have adopted [packaging conventions](https://packaging.python.org/).

Julia has a mature software packaging system and for publication VTUFileHandler.jl should take full advantage of it. I will let the author decide between the two options above, but one of them must be chosen as a requirement of publication. The choice will also affect how to integrate the XMLParser module.
