### 2022-04-13 #4300 @baxmittens
> Regarding package registration, I do not see how dependence on the other packages or lack there of would prevent package registration. This can always be modified in subsequent vesions of the package. VTUFileHandler.jl is either a reusable library or a terminal application.
>
> 1. A reusable library should be in a registry of some sort to facilitate installation.
> 2. Alternatively, a terminal application should have a Manifest.toml for reproducibility as well as a DOI via a service like Zenodo.
>
> Per the review criteria on https://joss.readthedocs.io/en/latest/review_criteria.html#installation-instructions:
>
> > Where possible, software installation should be managed with a package manager. However, this criterion depends upon the maturity and availability of software packaging and distribution in the programming language being used. For example, Python packages should be pip installable, and have adopted [packaging conventions](https://packaging.python.org/).
>
> Julia has a mature software packaging system and for publication VTUFileHandler.jl should take full advantage of it. I will let the author decide between the two options above, but one of them must be chosen as a requirement of publication. The choice will also affect how to integrate the XMLParser module.
>
> See https://github.com/JuliaRegistries/General#should-i-register-my-package

Sorry but I disagree, here.
My package already meets the review criteria (Good: The software is simple to install, and follows established distribution and dependency management approaches for the language being used).

Via

```julia
import Pkg
Pkg.add(url="https://github.com/baxmittens/XMLParser.git")
Pkg.add(url="https://github.com/baxmittens/VTUFileHandler.jl.git")
```

the package is registered in the julia package manager. it will be automatically updated if Pkg.update() is called and a new version is available. All dependencies are listed in the `Project.toml`.

A `Manifest.toml` is not needed here. You can think of the Manifest as a fixed snapshot of the Project.toml which is needed if your package requires specific versions/releases of its dependecies for a given application. In other words for specific applications it can be necessary that the user has the very same environment and thus a Manifest.toml has to be included.

This is not necessary here since my package should work with all versions of julia 1.x and only uses Julia.Base besides the XMLParser. I always thought that in that case you should avoid to add a Manifest.toml since this would result in overly restricting the versions of the package dependencies and leads to large `.julia` folders.

In some cases a `test/Manifest.toml` seems appropriate but in my case I don't see why. Only in case incompatibilites with other packages are occuring, I should add a Manifest.toml. Right?

I always thought this would be good julia practice. See for example [timholy's projects](https://github.com/timholy/ProgressMeter.jl).
