### 2022-04-13 #4300 @baxmittens
> > The problem is, to my knowledge, there is no XMLParser available which is written purely in Julia. All projects I know (LightXML.jl , EzXML.jl) use [libxml2](http://www.xmlsoft.org/) as a basis, which I wanted to avoid.
>
> It would be good to document this in the README for XMLParser.

Thank you for the suggestion. Added it in the readme.
