### 2022-04-13 #4300 @baxmittens
> > > 7. While the author asserts that this library "implements an algebra" I would like the author to defend this statement by addressing a few algebraic concepts.
> > > i) Are the operations associative?
> > > ii) Are the operations commutative?
> > > iii) Are there additive and multiplicative identities?
> > > iv) Are there additive and multiplicative inverses?
> > > v) Can multiplication be distributed over addition?
> > > vi) Over the set of `VTUFile`s do addition and multiplication form a [mathematical field](https://en.wikipedia.org/wiki/Field_(mathematics))?
> > > v) Over the set of `VTUFile`s and `Number`s do addition and multiplication form a multiplicative field?
> >
> >
> > As mentioned in the paper, if operators should be applied the objects have to share the same topology. If they do, all operators act _point-wise_ on each corresponding pair of e.g. nodal coefficients. The operators itself keep their properties as they perform their action on a collection of scalars. Each nodal coefficient $\hat{u}\in\mathbb{R}$ is an independet entity.
> > >
>
> Thank you for this comment. Please clarify the above points in the paper, clearly delineating how this library implements an algebra such as a field. If you feel there is an alternative algebra, please state what it is. Specifically there are a couple missing parts I have identified so far:

Let V=(VTUFile,+,*) be a field and A \subseteq  R^n a vector space over V. Then V is an algebra if for all x,y,z \in A and a,b \in V the following holds

1. (x+y) * z = x * z + y * z
2. z * (x+y) = Z * x + z * y
3. (ax) * (bx) = (ab)(x * y)

which surely applies but I don't really think this should be included in the paper or should it?

> 1. Additive inverse with a unary `-` operator.
> 2. Multiplicative identity (e.g. `Base.one`).

Added those, thank you.
