### 2022-04-13 #4300 @baxmittens
> The language of the title is very specific. So yes, I think the contents of the a article should support the statement in the title.

I would also consider changing the title to something like

`A VTU library in the Julia language that implements algebraic operators on VTU data`

if the language is too specific otherwise (I'm no native speaker).

We can also stick to the title and I define an algebra in the paper if this is the better solution
