### 2022-04-13 #4300 @mkitti
> You can think of the Manifest as a fixed snapshot of the Project.toml which is needed if your package requires specific versions/releases of its dependecies for a given application. In other words for specific applications it can be necessary that the user has the very same environment and thus a Manifest.toml has to be included.

Exactly. You have submitted this software for publication, and anyone reading the article should be able to reproduce the environment down to the specific versions of all the dependencies. This will actually simplify installation because the source of the XMLParser package will be encoded in the Manifest.toml:

```toml
[[deps.XMLParser]]
git-tree-sha1 = "5ffa66df38a330a7be8584dd8058e6f07fac928c"
repo-rev = "main"
repo-url = "https://github.com/baxmittens/XMLParser.jl.git"
uuid = "b09a1ce3-015a-45b4-89ae-f27efbd2c76d"
version = "0.0.4"
```

You could consider putting this on a specific branch related to this article. If this article is published, you could then add a tag at the time for publication, and then use Zenodo to acquire a permanent snapshot of the software at that tag which will have an associated DOI.

An application of the Manifest.toml is reproducibility for scientific publication. A second application is to capture unregistered dependencies. Thus, I believe a Manifest.toml is warranted here.

I would accept registration of the packages in lieu of this, but there is actually nothing stopping you from doing both. Your repository can have multiple branches. For the branch that you register, I would probably not include a Manifest.toml so that it can be easily combined with future versions of Julia and other packages.
