### 2022-04-13 #4300 @mkitti
# Documentation

The documentation currently consists of a single README.md. It currently has the following sections:
* Introduction
* Install
* Preliminaries
* Usage
* Example

The package overall needs more comprehensive documentation, and I recommend the use of https://github.com/JuliaDocs/Documenter.jl for this. The author may want to use the template located at https://github.com/invenia/PkgTemplates.jl/tree/master/templates/docs to get started.

The following sections need improvement:

## Statement of Need

This is partially addressed in the introduction. Perhaps it should borrow more from the corresponding section in the paper. A discussion of what this package does beyond ReadVTK.jl and WriteVTK.jl is needed as discussed above and why that functionality is needed.

## Installation instructions

The installation instructions work but as detailed above they could streamlined to meet the standards for software published in JOSS given the existing packaging framework in Julia.

## Example usage

The example includes a misspelling for "uncrompress".

A link to vox8.vtu in the `test` folder would be useful.

It currently covers some very simple arithmetic operations, mostly with scalars. The example should be expanded to include more operations between instances of `VTUFile`.

Ideally, an additional real world example beyond a simple cube would be included.

## Functionality documentation

Comprehensive API documentation is absent from the README. The code itself does contain doc strings, but due to to an extra space between the docstrings and the identifier, the doc strings are not correctly associated.

For example, the following line should be removed. Alternatively, `VTUHeader` could be placed there to associate the doc string.
https://github.com/baxmittens/VTUFileHandler.jl/blob/3aa96c4b15f155de4b73d826bcf0fa4e30f79e2b/src/VTUFileHandler.jl#L17

```
help?> VTUHeader
search:

Couldn't find VTUHeader
Perhaps you meant Threads
  No documentation found.

  Binding VTUHeader does not exist.
```

Documenter.jl would be easiest way to export the docstrings into readable documentation.

## Automated testing

There are automated tests but they are not described in the documentation. Additional code coverage is needed.

## Community guidelines

There are no community guidelines in the documentation.

> There should be clear guidelines for third-parties wishing to:
>
> Contribute to the software
> Report issues or problems with the software
> Seek support
>
