### 2022-04-13 #4300 @mkitti
# Paper

## Summary

There is not a clear description what a VTU file is other than naming it as "VTK unstructured grid". An introduction to this would be useful to make this paper more accessible to a non-specialist audience.

The author also leads with the motivation about "discrete numerical simulations". The software is likely has more general applicability than "discrete numerical simulations". Perhaps this description could be made to appeal to a wider audience.

# Statement of need

A statement of need section does exist in the paper
 A statement of need: Does the paper have a section titled 'Statement of Need' that clearly states what problems the software is designed to solve and who the target audience is?

# State of the field

This section is absent. There appears to be other `VTK` related I/O packages available and those should be discussed.

# Quality of the Writing

The language itself is good, however the structure of the writing seems to dive into the details of a specific application before addressing what the package actual does. For a non-specialist audience would be good to elaborate on some of the basics first:

1) What is VTK?
2) What is "VTK unstructured grid data"?
3) How do I plot "VTK unstructured grid data"?
4) What is the prior work in the field for similar packages in Julia and in other languages? Similar packages in Python are only mentioned in the very last sentence of the conclusion.
5) What is Julia?
6) What does this package allow the user to do?

What is currently in the Introduction perhaps belongs in a Motivation section in a subsequent section after a basic introduction.

# References

VTK itself is not properly cited as [requested](https://vtk.org/about/).

> Schroeder, Will; Martin, Ken; Lorensen, Bill (2006), The Visualization Toolkit (4th ed.), Kitware, [ISBN](http://en.wikipedia.org/wiki/International_Standard_Book_Number) [978-1-930934-19-1](http://en.wikipedia.org/wiki/Special:BookSources/978-1-930934-19-1)
