### 2022-04-13 #4300 @mkitti
`VTUData` and several other struct types are declared to be `mutable` but it is not clear to me that they need to be. All of `VTUData`'s fields are `Vector`s, which themselves are mutable. Why is `VTUData` mutable?
