### 2022-04-14 #4300 @jbytecode
a small note for test:

The Test package includes useful macros for testing functions provided by the package and a standard testing scheme is

```
@testset "Test Description" begin
      // Some calculations
     result = somefunction(...)
     @test expected == result
end
```

in Julia. The authors may consider refactoring their test files using the standard Julia scheme.
