### 2022-04-22 #4300 @baxmittens
@mkitti - tried to adapt the paper according to your suggestions.
