### 2022-04-22 #4300 @mkitti
I suggest asking https://discourse.julialang.org for the items you are struggling with.

The chat resources here can provide some interactive feedback as well:
https://julialang.org/community/

Package registration can be done via a web form if preferred:
https://github.com/JuliaRegistries/Registrator.jl

To get started I usually would use https://github.com/invenia/PkgTemplates.jl in interactive mode.
