### 2022-05-10 #4300 @baxmittens
@jbytecode - Following mkitti's and your recommendations, I have registered both GitHub projects as Julia packages and added documentations. I  added a section to the paper and did some alterations to the code.
I have to write some unit test now, define a test enviroment and do some minor things here and there and then I think I will be finished.
This will take me until next week, I guess.
