### 2022-05-13 #4300 @baxmittens
@mkitti - I added a CI.yml and a branch `joss_paper` with a `Manifest.toml`. Is this sufficient or need I to include Zenodo as well?
And if so, how is it done? Do I upload a zip-file of this branch? I'm not quite sure.
