### 2022-05-13 #4300 @jbytecode
You don't need a Zenodo archive in this stage.
