### 2022-05-13 #4300 @mkitti
Package registration satisfies my prior concerns. The Manifest.toml is appreciated. Zenodo would be most applicable at publication time.
