### 2022-05-17 #4300 @mkitti
Overall, the paper, software, and documentation have greatly improved during the course of this review. The submitted paper describes a valuable tool to process VTU files from the VTK framework. It provides arithmetic operations on the types representing the files that complete an algebraic field. Additionally, a supporting package has an accessory package, XMLParser.jl that provides a lightweight XML framework. The software is easy to install. The paper introduces the background to a general audience and describes the main functionality of the paper. I thus recommend the paper for publication.

I have listed a few minor points below that I believe can be easily addressed. I may a few additional minor recommendations after I do a close reading of the paper later this week.

# Minor Points

## 1. Code or command to generate Figure 1

I would appreciate if the author could provide code to generate Figure 1 or at least view the VTU file in VTK. This could also be a command line invocation on the VTU file.

## 2. Community Guidelines

The current community guidelines listed are as follows:

> Contributions to or questions about this project are welcome. Feel free to create a issue or a pull request.

I would clarify that you would like them to create an issue or pull request on Github.

It's possible that the README may be read on other websites other than Github. For example:
https://juliahub.com/ui/Packages/VTUFileHandler/udkbV/0.1.3

## 3. Link back to the repository from the documentation

Similar to the above, also please provide a link from the documentation to the Github repository since the documentation may be read on sites other than Github or even offline.

https://docs.juliahub.com/VTUFileHandler/udkbV/0.1.3/

## 4. Document implemented base methods

I would appreciate a brief listing of what `Base` methods are implemented for `VTUFile` in the documentation would be useful.

An incomplete list is as follows:
* `Base.write`
* `Base.deepcopy`
* `Base.similar`
* `Base.fill!`
* `Base.zero`
* `Base.one`
* `Base.empty!`
* `Base.empty`
* `Base.getindex`
