### 2022-05-18 #4300 @baxmittens
@mkitti -

"""
1. Code or command to generate Figure 1
"""
I do not use any code to render the images. maybe this is possible with some additional vtk-software. I just use paraview and insert the files manually. i added that to the paper. Is that ok?

I think I have implemented the rest of the aforementioned demands.
