### 2022-05-18 #4300 @mkitti
Do you just use the Paraview UI to File -> Open the VTU files?

If I typed the following in a shell, would it open the files?

```
paraview vox_8.vtu vox_8_1.vtu
```
