### 2022-05-18 #4300 @baxmittens
> Do you just use the Paraview UI to File -> Open the VTU files?
>
> If I typed the following in a shell, would it open the files?
>
> ```
> paraview vox8.vtu vox8_1.vtu
> ```

I have Paraview installed on windows and I couldn't figure out how to load several files by command line at once

`paraview vox8.vtu`

or

`paraview --data=vox8.vtu`

works for me but not for two files

[In the docs](https://docs.paraview.org/en/latest/UsersGuide/dataIngestion.html) there is the data-option defined an a additional option for a (time)-series but not for two seperate inputs...at least as far as I'm aware
