### 2022-05-19 #4300 @dmbates
> > > The problem is, to my knowledge, there is no XMLParser available which is written purely in Julia. All projects I know (LightXML.jl , EzXML.jl) use [libxml2](http://www.xmlsoft.org/) as a basis, which I wanted to avoid.
> >
> >
> > It would be good to document this in the README for XMLParser.
>
> Thank you for the suggestion. Added it in the readme.

First, my thanks for @mkitti for his thoughtful and comprehensive review through many iterations.  I have more-or-less just monitored the conversation because @mkitti was bringing up the points that I would have made (and many more).

I am grateful to see the adoption of a registered package.  Part of the conversation seemed to be trying to avoid using the Julia package system to its fullest extent. That would be unfortunate if it results in reinventing the wheel.

The above comment regarding not using a package that depends on libxml2 is an example.  Package systems for other languages would require shipping the source code for a library like libxml2 and working out the onerous details of compiling and linking on systems to which the author would not have access. That is a big pain and can become a big time sink.  However, the BinaryBuilder.jl framework circumvents this process in Julia.  You can add a dependency on the [XML2_jll package](https://juliahub.com/ui/Packages/XML2_jll/xWVg2/2.9.12+0?page=2) and the nasty details of compiling and linking are handled for you.  This allows you to access well designed and tested code without needing to be concerned with the provision of source code, configuration, compilation and linkage details.

In general there will be a trade-off between learning to use code provided in a library and building and testing code in Julia with similar capabilities.  The purpose of BinaryBuilder.jl is to make it easier to do the former.
