### 2022-05-19 #4300 @mkitti
> The above comment regarding not using a package that depends on libxml2 is an example. Package systems for other languages would require shipping the source code for a library like libxml2 and working out the onerous details of compiling and linking on systems to which the author would not have access. That is a big pain and can become a big time sink. However, the BinaryBuilder.jl framework circumvents this process in Julia. You can add a dependency on the [XML2_jll package](https://juliahub.com/ui/Packages/XML2_jll/xWVg2/2.9.12+0?page=2) and the nasty details of compiling and linking are handled for you. This allows you to access well designed and tested code without needing to be concerned with the provision of source code, configuration, compilation and linkage details.

There is some value in having a lightweight XML package that does not have an external binary dependency. It would be great if they could coordinate by implementing a common "AbstractXMLParser" interface in the same fashion as https://github.com/JuliaMath/AbstractFFTs.jl.

In this case, it might be worthwhile to combine functionality into https://github.com/JuliaComputing/XML.jl (recently relocated), but this exceeds the scope of the review.
