### 2022-05-23 #4300 @jbytecode
- line 26 : julia -> Julia
- line 41 : remove the line =========

@baxmittens could you please fix those items in paper? please have a full read and correct other typos if exist.
