### 2022-05-24 #4300 @jbytecode
@baxmittens thank you for the submission, @mkitti and @dmbates thank you for your valuable reviews and consuming your time for JOSS.

One of the @openjournals/joss-eics will have the final decision.

Thank you!
