### 2022-05-24 #4300 @baxmittens
@jbytecode @mkitti  @dmbates  - thank you for reviewing my first joss article.
Your help greatly improved the quality of my project and the paper as well. I really appriciate it!
I will publish further articels here, i guess. The next time I will be better prepared and setting up documentaion and stuff will only take me a fraction of the time I spent for this project. So I really feel I do benefit from undergoing this review procedure. Thanks again to you all.
